test(test262): merge assignment execution test files
